Fix more log format warning
The expression port + 1 is promoted to int, but printed as uint16_t.

This is the same mistake fixed for a different log by
7eb16ce36458011d87972715f08bd9f03ff39807.

Refs #1726 <https://github.com/Genymobile/scrcpy/issues/1726>

+++ app/src/server.c
@@ -201,7 +201,7 @@ enable_tunnel_forward_any_port(struct server *server,
 
         if (port < port_range.last) {
             LOGW("Could not forward port %" PRIu16", retrying on %" PRIu16,
-                 port, port + 1);
+                 port, (uint16_t) (port + 1));
             port++;
             continue;
         }
